requirements.txt
Django==2.2.3
djangorestframework==3.10.1
dj-database-url==0.5.0
dj-static==0.0.6
python-decouple==3.1
static3==0.7.0
easy-thumbnails==2.6
django-compressor==2.3
django-widget-tweaks==1.4.5
django-embed-video==1.2.0
googlemaps==3.0.2
django-tinymce4-lite==1.7.5
django-filebrowser-no-grappelli==3.7.8
django-crispy-forms==1.7.2
django-active-link==0.1.6

#dev
bpython==0.18
django-extensions==2.1.9

#prod
gunicorn==19.8.1